refactor(fakeNitro): extract into separate files
for better maintainability and readability
