+++ src/plugins/fakeNitro/regexes.ts
@@ -1,0 +1,10 @@
+/*
+ * Vencord, a Discord client mod
+ * Copyright (c) 2024 Vendicated and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+export const fakeNitroEmojiRegex = /\/emojis\/(\d+?)\.(png|webp|gif)/;
+export const fakeNitroStickerRegex = /\/stickers\/(\d+?)\./;
+export const fakeNitroGifStickerRegex = /\/attachments\/\d+?\/\d+?\/(\d+?)\.gif/;
+export const hyperLinkRegex = /\[.+?\]\((https?:\/\/.+?)\)/;
